server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.0
beanie==1.25.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
